Minimal example of DUB dependency clashes
This is a simple example of the dependency resolution failure described
in https://github.com/dlang/dub/issues/1732.  Running `dub build` fails
with the error:

    Unresolvable dependencies to package taggedalgebraic:
      eventcore 0.8.43 depends on taggedalgebraic >=0.10.12 <0.12.0-0
      std_data_json 0.18.3 depends on taggedalgebraic ~>0.10.1

The reported failure in fact has an easily recognizable resolution, but
for some reason DUB is unable to find it.  It is not 100% clear if this
is the real dependency clash or not, but it is interesting to note that
if `dep2` (which has no dependencies of its own) is removed from the
list of dependencies of the main project, the resolution failure also
disappears.